+++ NFlags/Arguments/FlagBuilder.cs
@@ -1,0 +1,103 @@
+namespace NFlags.Arguments
+{
+    /// <summary>
+    /// Builder for flag arguments
+    /// </summary>
+    public class FlagBuilder
+    {
+        private string _name;
+        private string _description;
+        private string _abr;
+        private bool _isPersistent;
+        private bool _defaultValue;
+        private string _environmentVariable;
+
+        /// <summary>
+        /// Set name of the argument.
+        /// </summary>
+        /// <param name="name">Name of the argument</param>
+        /// <returns>Self instance</returns>
+        public FlagBuilder Name(string name)
+        {
+            _name = name;
+
+            return this;
+        }
+
+        /// <summary>
+        /// Set description of the argument.
+        /// </summary>
+        /// <param name="description">Description of the argument</param>
+        /// <returns>Self instance</returns>
+        public FlagBuilder Description(string description)
+        {
+            _description = description;
+
+            return this;
+        }
+
+        /// <summary>
+        /// Set flag abbreviation
+        /// </summary>
+        /// <param name="abr">Flag abbreviation</param>
+        /// <returns>Self instance</returns>
+        public FlagBuilder Abr(string abr)
+        {
+            _abr = abr;
+
+            return this;
+        }
+
+        /// <summary>
+        /// Make flag persistent.
+        /// </summary>
+        /// <returns>Self instance</returns>
+        public FlagBuilder Persistent()
+        {
+            _isPersistent = true;
+
+            return this;
+        }
+
+        /// <summary>
+        /// Set default value of the argument
+        /// </summary>
+        /// <param name="defaultValue">Default value of the argument</param>
+        /// <returns>Self instance</returns>
+        public FlagBuilder DefaultValue(bool defaultValue)
+        {
+            _defaultValue = defaultValue;
+
+            return this;
+        }
+
+        /// <summary>
+        /// Set environment variable name for the argument value
+        /// </summary>
+        /// <param name="environmentVariable">Environment variable name for the argument value</param>
+        /// <returns>Self instance</returns>
+        public FlagBuilder EnvironmentVariable(string environmentVariable)
+        {
+            _environmentVariable = environmentVariable;
+
+            return this;
+        }
+
+        /// <summary>
+        /// Build flag
+        /// </summary>
+        /// <returns>Flag</returns>
+        public Flag Build()
+        {
+            return new Flag
+            {
+                Name = _name,
+                Description = _description,
+                DefaultValue = _defaultValue,
+                EnvironmentVariable = _environmentVariable,
+                Abr = _abr,
+                IsPersistent = _isPersistent
+            };
+        }
+    }
+}